refactor: add constants and typings for LSP2
