+++ packages/lsp2-contracts/constants.ts
@@ -1,0 +1,7 @@
+export type Verification = {
+  method: string;
+  data: string;
+  source?: string;
+};
+
+export type LSP2ArrayKey = { length: string; index: string };
